- Frames per video = 17 - Change data preprocessing: now we save frames from video only if we could find exactly frames_per_video faces

+++ src/prepare_data.py
@@ -1,16 +1,15 @@
 import os
 import json
 from multiprocessing import Pool
-from typing import List, Tuple
+from typing import List, Tuple, Union
 
 import cv2
-import cvlib as cv
 import numpy as np
 from sklearn.model_selection import GroupShuffleSplit
 from tqdm import tqdm
 
 import config
-from utils import extend_rect_to_square, read_faces_from_video
+from utils import read_faces_from_video
 
 
 def get_train_data_lists() -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
@@ -27,7 +26,7 @@ def get_train_data_lists() -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
 
         metadata_file_path = os.path.join(chunk_dir, 'metadata.json')
         with open(metadata_file_path) as json_file:
-            chunk_metadata = json.load(json_file
+            chunk_metadata = json.load(json_file)
 
         for video_file_name in chunk_file_names:
             paths.append(os.path.join(chunk_dir, video_file_name))
@@ -40,55 +39,81 @@ def get_train_data_lists() -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
 
 def get_image_name_from_video_name(video_name, frame_number=None):
     if frame_number is not None:
-        return video_name[:-4] + '_{}.png'.format(frame_number)
+        if frame_number > 99:
+            template = '_{}.png'
+        elif 9 < frame_number < 100:
+            template = '_0{}.png'
+        else:
+            template = '_00{}.png'
+        return video_name[:-4] + template.format(frame_number)
     else:
         return video_name[:-4] + '.png'
 
 
-def save_faces_from_video(path: str) -> List[str]:
-    video_name = os.path.basename(path)
-    image_paths = []
-    for i in range(FRAMES_PER_VIDEO):
+def save_faces_from_video(video_path: str) -> Union[None, Tuple]:
+    video_name = os.path.basename(video_path)
+    face_paths = []
+    group = video_path.split('/')[-2]
+    for i in range(config.FRAMES_PER_VIDEO):
         image_name = get_image_name_from_video_name(video_name, i)
-        image_paths.append(os.path.join(config.TRAIN_FACES_DIR, image_name))
+        face_paths.append(os.path.join(config.TRAIN_FACES_DIR, image_name))
 
-    for image_path in image_paths:
-        if os.path.exists(image_path):
-            return image_paths
+    all_exists = True
+    for image_path in face_paths:
+        if not os.path.exists(image_path):
+            all_exists = False
+            break
+    if all_exists:
+        return face_paths, video_path, group
 
     try:
-        faces_to_save = read_faces_from_video(path, swap_channels=False)
+        faces_to_save = read_faces_from_video(video_path, swap_channels=False)
     except Exception:
-            print('Cannot read video:', path)
+        return video_path
 
-    for image_path, face in zip(image_paths, faces_to_save):
+    for image_path, face in zip(face_paths, faces_to_save):
         try:
             cv2.imwrite(image_path, face)
         except Exception:
-            print('Cannot save face image:', path, face.shape)
-    return image_paths
+            return video_path
+    return face_paths, video_path, group
 
 
 def extract_faces_from_videos_parallel(paths: List[str]) -> np.array:
-    os.makedirs(TRAIN_FACES_DIR, exist_ok=True)
-
-    faces = []
-    with Pool() as p:
+    os.makedirs(config.TRAIN_FACES_DIR, exist_ok=True)
+    faces_metadata = []
+    broken_videos = []
+    with Pool(20) as p:
         with tqdm(total=len(paths)) as pbar:
-            for i, video_faces_paths in enumerate(p.imap_unordered(save_faces_from_video, paths)):
-                faces.extend(video_faces_paths)
+            for i, saving_result in enumerate(p.imap_unordered(save_faces_from_video, paths)):
+                # print('saving_result:', saving_result)
+                # print('type(saving_result):', type(saving_result))
+                if type(saving_result) is tuple:
+                    faces_metadata.append({
+                        'faces': saving_result[0],
+                        'video': os.path.basename(saving_result[1]),
+                        'group': saving_result[2]
+                    })
+                else:
+                    broken_videos.append(saving_result)
                 pbar.update()
-    return np.array(faces)
+
+    print('{} broken files out of {}'.format(len(broken_videos), len(paths)))
+
+    with open(config.BROKEN_VIDEOS_PATH, 'w') as outfile:
+        json.dump(broken_videos, outfile)
+
+    return np.array(faces_metadata)
 
 
-def save_faces_metadata_to_json(video_paths, labels, groups, face_paths, metadata_path) -> None:
+def save_faces_metadata_to_json(labels, faces, metadata_path) -> None:
     faces_metadata = {}
-    for video_path, label, group, face_paths_batch in zip(video_paths, labels, groups, face_paths):
-        for face_path in face_paths_batch:
+    for label, faces_batch in zip(labels, faces):
+        for face_path in faces_batch['faces']:
             faces_metadata[os.path.basename(face_path)] = {
                 'label': int(label),
-                'group': group,
-                'video': os.path.basename(video_path)
+                'group': faces_batch['group'],
+                'video': faces_batch['video']
             }
 
     with open(metadata_path, 'w') as outfile:
@@ -98,36 +123,31 @@ def save_faces_metadata_to_json(video_paths, labels, groups, face_paths, metadat
 if __name__ == '__main__':
     np.random.seed(config.SEED)
     video_paths, video_file_names, labels, groups = get_train_data_lists()
+
+    video_paths, video_file_names, labels, groups = video_paths, video_file_names, labels, groups
+
     print('Total videos:', video_paths.shape)
 
     print('Extracting faces from videos ...')
-    face_paths = extract_faces_from_videos_parallel(video_paths)
+    faces = extract_faces_from_videos_parallel(video_paths)
 
     print('Splitting data on train and validation ...')
-    order_index = np.argsort(video_file_names)
-    video_paths = video_paths[order_index]
-    labels = labels[order_index]
-    groups = groups[order_index]
-    face_paths = np.sort(face_paths)
-
-    face_paths = np.reshape(face_paths, (video_paths.shape[0], config.FRAMES_PER_VIDEO))
-
-    gss = GroupShuffleSplit(n_splits=1, train_size=.95, random_state=config.SEED)
-    train_idx, val_idx = next(gss.split(video_paths, labels, groups))
+    faces_groups = [face['group'] for face in faces]
+    faces_videos = [face['video'] for face in faces]
+    videos_labels_dict = dict(zip(video_file_names, labels))
+    faces_labels = np.array([videos_labels_dict[video] for video in faces_videos])
+    gss = GroupShuffleSplit(n_splits=1, train_size=0.8, random_state=config.SEED)
+    train_idx, val_idx = next(gss.split(faces, faces_labels, faces_groups))
     print('Train len:', len(train_idx), ', Val len:', len(val_idx))
 
     print('Saving faces metadata ...')
     save_faces_metadata_to_json(
-        video_paths[train_idx],
-        labels[train_idx],
-        groups[train_idx],
-        face_paths[train_idx],
+        faces_labels[train_idx],
+        faces[train_idx],
         config.FACES_TRAIN_METADATA_PATH)
     save_faces_metadata_to_json(
-        video_paths[val_idx],
-        labels[val_idx],
-        groups[val_idx],
-        face_paths[val_idx],
+        faces_labels[val_idx],
+        faces[val_idx],
         config.FACES_VAL_METADATA_PATH)
 
     print('All done!')
